requirements.txt
pip
pillow
numpy
numba
argparse
tabulate